+++ core/store/orm/config.go
@@ -82,8 +82,10 @@ type (
 		EthGasBumpWei                    *big.Int
 		EthGasPriceDefault               *big.Int
 		EthMaxGasPriceWei                *big.Int
+		EthMinGasPriceWei                *big.Int
 		EthFinalityDepth                 uint
 		EthHeadTrackerHistoryDepth       uint
+		EthHeadTrackerSamplingInterval   time.Duration
 		EthBalanceMonitorBlockDelay      uint16
 		EthTxResendAfterThreshold        time.Duration
 		GasUpdaterBatchSize              *uint32
@@ -108,8 +110,10 @@ func init() {
 		EthGasBumpWei:                    big.NewInt(5000000000),    // 5 Gwei
 		EthGasPriceDefault:               big.NewInt(20000000000),   // 20 Gwei
 		EthMaxGasPriceWei:                big.NewInt(1500000000000), // 1.5 Twei
+		EthMinGasPriceWei:                big.NewInt(1000000000),    // 1 Gwei
 		EthFinalityDepth:                 50,
 		EthHeadTrackerHistoryDepth:       100,
+		EthHeadTrackerSamplingInterval:   1 * time.Second,
 		EthBalanceMonitorBlockDelay:      1,
 		EthTxResendAfterThreshold:        30 * time.Second,
 		GasUpdaterBlockDelay:             1,
@@ -123,9 +127,23 @@ func init() {
 
 	// NOTE: There are probably other variables we can tweak for Kovan and other
 	// test chains, but requires more in-depth research on their consensus
-	// mechanisms. For now, mainnet defaults ought to be safe
+	// mechanisms. For now, mainnet defaults ought to be safe enough for testnet.
 	kovan := mainnet
 	kovan.HeadTimeBudget = 4 * time.Second
+
+	// xDai currently uses AuRa (like Parity) consensus so finality rules will be similar to parity
+	// See: https://www.poa.network/for-users/whitepaper/poadao-v1/proof-of-authority
+	// NOTE: xDai is planning to move to Honeybadger BFT which might have different finality guarantees
+	// https://www.xdaichain.com/for-validators/consensus/honeybadger-bft-consensus
+	// For worst case re-org depth on AuRa, assume 2n+2 (see: https://github.com/poanetwork/wiki/wiki/Aura-Consensus-Protocol-Audit)
+	// With xDai's current maximum of 19 validators then 40 blocks is the maximum possible re-org)
+	// The mainnet default of 50 blocks is ok here
+	xDai := mainnet
+	xDai.EthGasBumpThreshold = 8                      // mainnet * 2.8 ish (5s vs 13s block time)
+	xDai.EthGasPriceDefault = big.NewInt(1000000000)  // 1 Gwei
+	xDai.EthMinGasPriceWei = big.NewInt(1000000000)   // 1 Gwei is the minimum accepted by the validators (unless whitelisted)
+	xDai.EthMaxGasPriceWei = big.NewInt(500000000000) // 500 Gwei
+	xDai.HeadTimeBudget = 5 * time.Second
 
 	// BSC uses Clique consensus with ~3s block times
 	// Clique offers finality within (N/2)+1 blocks where N is number of signers
@@ -136,8 +154,9 @@ func init() {
 		EthGasBumpWei:                    big.NewInt(5000000000),   // 5 Gwei
 		EthGasPriceDefault:               big.NewInt(5000000000),   // 5 Gwei
 		EthMaxGasPriceWei:                big.NewInt(500000000000), // 500 Gwei
-		EthFinalityDepth:                 50,                       // Keeping this > 11 because it's not expensive and gives us a safety margin
+		EthFinalityDepth:                 50,                       // Keeping this >> 11 because it's not expensive and gives us a safety margin
 		EthHeadTrackerHistoryDepth:       100,
+		EthHeadTrackerSamplingInterval:   1 * time.Second,
 		EthBalanceMonitorBlockDelay:      2,
 		EthTxResendAfterThreshold:        15 * time.Second,
 		GasUpdaterBlockDelay:             2,
@@ -160,9 +179,10 @@ func init() {
 		EthMaxGasPriceWei:                big.NewInt(500000000000), // 500 Gwei
 		EthFinalityDepth:                 200,                      // A sprint is 64 blocks long and doesn't guarantee finality. To be safe, we take three sprints (192 blocks) plus a safety margin
 		EthHeadTrackerHistoryDepth:       250,                      // EthFinalityDepth + safety margin
-		EthBalanceMonitorBlockDelay:      13,                       // equivalent of 1 eth block seems reasonable
-		EthTxResendAfterThreshold:        5 * time.Minute,          // 5 minutes is roughly 300 blocks on Matic. Since re-orgs occur often and can be deep, we want to avoid overloading the node with a ton of re-sent unconfirmed transactions.
-		GasUpdaterBlockDelay:             32,                       // Delay needs to be large on matic since re-orgs are so frequent at the top level
+		EthHeadTrackerSamplingInterval:   1 * time.Second,
+		EthBalanceMonitorBlockDelay:      13,              // equivalent of 1 eth block seems reasonable
+		EthTxResendAfterThreshold:        5 * time.Minute, // 5 minutes is roughly 300 blocks on Matic. Since re-orgs occur often and can be deep, we want to avoid overloading the node with a ton of re-sent unconfirmed transactions.
+		GasUpdaterBlockDelay:             32,              // Delay needs to be large on matic since re-orgs are so frequent at the top level
 		GasUpdaterBlockHistorySize:       128,
 		GasUpdaterBatchSize:              &defaultGasUpdaterBatchSize,
 		GasUpdaterEnabled:                true,
@@ -177,6 +197,7 @@ func init() {
 		EthGasBumpThreshold:              0, // Never bump gas on optimism
 		EthFinalityDepth:                 1, // Sequencer offers absolute finality as long as no re-org longer than 20 blocks occurs on main chain, this event would require special handling (new txm)
 		EthHeadTrackerHistoryDepth:       10,
+		EthHeadTrackerSamplingInterval:   1 * time.Second,
 		EthBalanceMonitorBlockDelay:      0,
 		EthTxResendAfterThreshold:        5 * time.Second,
 		GasUpdaterBlockHistorySize:       0, // Force an error if someone set GAS_UPDATER_ENABLED=true by accident; we never want to run the gas updater on optimism
@@ -198,6 +219,8 @@ func init() {
 	ChainSpecificDefaults[56] = bscMainnet
 	ChainSpecificDefaults[128] = hecoMainnet
 	ChainSpecificDefaults[80001] = polygonMatic
+
+	ChainSpecificDefaults[100] = xDai
 }
 
 func chainSpecificConfig(c Config) ChainSpecificDefaultSet {
@@ -257,6 +280,13 @@ func (c *Config) Validate() error {
 			c.EthGasBumpPercent(),
 			ethCore.DefaultTxPoolConfig.PriceBump,
 		)
+	}
+
+	if c.EthMinGasPriceWei().Cmp(c.EthGasPriceDefault()) > 0 {
+		return errors.New("ETH_MIN_GAS_PRICE_WEI must be less than or equal to ETH_GAS_PRICE_DEFAULT")
+	}
+	if c.EthMaxGasPriceWei().Cmp(c.EthGasPriceDefault()) < 0 {
+		return errors.New("ETH_MAX_GAS_PRICE_WEI must be greater than or equal to ETH_GAS_PRICE_DEFAULT")
 	}
 
 	if c.EthHeadTrackerHistoryDepth() < c.EthFinalityDepth() {
@@ -608,7 +638,6 @@ func (c Config) EthGasBumpWei() *big.Int {
 // EthMaxGasPriceWei is the maximum amount in Wei that a transaction will be
 // bumped to before abandoning it and marking it as errored.
 func (c Config) EthMaxGasPriceWei() *big.Int {
-
 	str := c.viper.GetString(EnvVarName("EthMaxGasPriceWei"))
 	if str != "" {
 		n, err := parseBigInt(str)
@@ -630,6 +659,24 @@ func (c Config) EthMaxGasPriceWei() *big.Int {
 // 0 value disables
 func (c Config) EthMaxUnconfirmedTransactions() uint64 {
 	return c.getWithFallback("EthMaxUnconfirmedTransactions", parseUint64).(uint64)
+}
+
+// EthMinGasPriceWei is the minimum amount in Wei that a transaction may be priced.
+// Chainlink will never send a transaction priced below this amount.
+func (c Config) EthMinGasPriceWei() *big.Int {
+	str := c.viper.GetString(EnvVarName("EthMinGasPriceWei"))
+	if str != "" {
+		n, err := parseBigInt(str)
+		if err != nil {
+			logger.Errorw(
+				"Invalid value provided for EthMinGasPriceWei, falling back to default.",
+				"value", str,
+				"error", err)
+		} else {
+			return n.(*big.Int)
+		}
+	}
+	return chainSpecificConfig(c).EthMinGasPriceWei
 }
 
 // EthNonceAutoSync enables/disables running the NonceSyncer on application start
@@ -679,6 +726,14 @@ func (c Config) EthGasLimitMultiplier() float32 {
 
 // SetEthGasPriceDefault saves a runtime value for the default gas price for transactions
 func (c Config) SetEthGasPriceDefault(value *big.Int) error {
+	min := c.EthMinGasPriceWei()
+	max := c.EthMaxGasPriceWei()
+	if value.Cmp(min) < 0 {
+		return errors.Errorf("cannot set default gas price to %s, it is below the minimum allowed value of %s", value.String(), min.String())
+	}
+	if value.Cmp(max) > 0 {
+		return errors.Errorf("cannot set default gas price to %s, it is above the maximum allowed value of %s", value.String(), max.String())
+	}
 	if c.runtimeStore == nil {
 		return errors.New("No runtime store installed")
 	}
@@ -729,6 +784,24 @@ func (c Config) EthHeadTrackerHistoryDepth() uint {
 // for the head tracker before we start dropping heads to keep up.
 func (c Config) EthHeadTrackerMaxBufferSize() uint {
 	return uint(c.getWithFallback("EthHeadTrackerMaxBufferSize", parseUint64).(uint64))
+}
+
+// EthHeadTrackerSamplingInterval is the interval between sampled head callbacks
+// to services that are only interested in the latest head every some time
+func (c Config) EthHeadTrackerSamplingInterval() time.Duration {
+	str := c.viper.GetString(EnvVarName("EthHeadTrackerSamplingInterval"))
+	if str != "" {
+		n, err := parseDuration(str)
+		if err != nil {
+			logger.Errorw(
+				"Invalid value provided for EthHeadTrackerSamplingInterval, falling back to default.",
+				"value", str,
+				"error", err)
+		} else {
+			return n.(time.Duration)
+		}
+	}
+	return chainSpecificConfig(c).EthHeadTrackerSamplingInterval
 }
 
 // EthTxResendAfterThreshold controls how long the ethResender will wait before